refactor(dave): Extract query-client to its own file to be a singleton.
