+++ apps/dave/src/providers/DataProvider.tsx
@@ -1,41 +1,11 @@
 import { CartesiProvider } from "@cartesi/wagmi";
-import {
-    QueryClient,
-    QueryClientProvider,
-    type QueryClientConfig,
-} from "@tanstack/react-query";
-import { defaultTo, pipe } from "ramda";
+import { QueryClientProvider } from "@tanstack/react-query";
 import type { FC, PropsWithChildren } from "react";
-import addTestCaseData from "./localdata/cases";
-
-const mockEnabled = defaultTo(
-    false,
-    process.env.NEXT_PUBLIC_MOCK_ENABLED === "true",
-);
+import queryClient from "./queryClient";
 
 const nodeRpcUrl =
     process.env.NEXT_PUBLIC_CARTESI_NODE_RPC_URL ??
     "http://127.0.0.1:10011/rpc";
-
-const injectData = pipe(addTestCaseData);
-
-const queryClientConfig: QueryClientConfig = mockEnabled
-    ? {
-          defaultOptions: {
-              queries: {
-                  staleTime: Infinity,
-                  gcTime: Infinity,
-                  networkMode: "always",
-              },
-          },
-      }
-    : {};
-
-const queryClient = new QueryClient(queryClientConfig);
-
-if (mockEnabled) {
-    injectData(queryClient);
-}
 
 const DataProvider: FC<PropsWithChildren> = ({ children }) => {
     return (
